BUZZ-191 - Move data type subscribing to separate file
